// Auto-Parts-Management-insurance_form/Client/angular-parthub/src/app/models/note.model.ts
export class Note {

  id?: any = 0;
  notes?: string;
  color?: string = 'text-white';
  colorVehicle?: string; //none bootstarp color
  userId?: number;
  employeeId?: any;
  jobId?: any;
  vehicleId?: any;
  type?: any;
  companyId?: number;
   
  year?:any;
  make?:any;
  model?:any;
  jobName?: any;
  jobPrice?: any;
  jobStatus?: any;
  sequenceNumber?: any = 0;

  createdAt?: Date;

  creatorShortName?: any;
}